Update to Bevy 0.15 release candidate (#540)
# Objective

The Bevy 0.15 release candidate has been released! Avian should be updated for early adopters, and to make sure we are ready when the full 0.15 release is out.

## Solution

Migrate Avian to the Bevy 0.15 release candidate. This only includes the migration necessary to fix all deprecations and warnings. Avian's internals will be updated to take advantage of required components and other new features in follow-ups.

---

## Migration Guide

`Collider::regular_polygon` and `ColliderConstructor::RegularPolygon` now use a `u32` instead of `usize` for `sides`.