Add physics picking backend using `bevy_picking` (#554)
# Objective

Bevy 0.15 upstreams `bevy_mod_picking` as `bevy_picking`. Now that picking is first-party, we should add a collider picking backend, especially as `bevy_mod_picking` (which has an Avian backend) is on track for deprecation.

## Solution

Add a `PhysicsPickingPlugin`! It is behind the `bevy_picking` default feature and must be added manually, but once it is enabled, picking is enabled for all colliders by default. Picking can be made opt-in using `PhysicsPickingSettings::require_markers`, in which case you must mark pickable entities with the `PhysicsPickable` component. The API and approach quite closely match the `MeshPickingPlugin` in Bevy 0.15.

The plugin is largely based on the [Avian backend](https://github.com/aevyrie/bevy_mod_picking/blob/7494b9614a50a9c25e77a9bdccdb9c010a506ae9/examples/avian.rs) in `bevy_mod_picking`, with some tweaks to docs and naming. I also added a 2D version of the picking backend.

There is a new `picking_3d` example:

https://github.com/user-attachments/assets/5fcac14e-c9ad-40b0-81a6-450fb32ec23b

(Note: The torus and conical frustum use trimesh colliders here, which is why the normals change a bit abruptly)

I have not added a 2D example (yet), but tested that it works:

https://github.com/user-attachments/assets/ec52c846-d4bc-4df5-bfb3-5b155d33492e