Make sure `Collider::triangle` is oriented CCW (#579)
# Objective

Fixes #368.

Many users create 2D triangle colliders with clockwise orientation. However, this produces a collider that doesn't work, as Parry seems to require conterclockwise orientation.

## Solution

Check the orientation in `Collider::triangle`, and reverse it if necessary. If the orientation is known to be counterclockwise, `Collider::triangle_unchecked` can be used instead.