Add `SolverSchedulePlugin` to encapsulate solver scheduling (#577)
# Objective

Fixes #571.

Currently, a lot of plugins depend on the system sets of the solver, even if they don't strictly need the solver. Most of these system sets are initialized by the `SolverPlugin`, and if you disable it, you get a crash.

We should aim to decouple system set ordering and the substepping loop scheduling from the actual solver, allowing the engine to rely on ordering guarantees while letting users freely disable and modify parts of the engine.

## Solution

Add a `SolverSchedulePlugin`, which sets up core system sets for the physics solver, and runs the substepping loop. Types related to the solver and substepping loop have been moved over from the `PhysicsSchedulePlugin`.

This isn't a perfect solution yet, and we could certainly decouple and reorganize things more, but it is a step forward.

---

## Migration Guide

System set configuration and scheduling related to the solver and substepping loop are now primarily in the new `SolverSchedulePlugin`. It is included in the `PhysicsPlugins` plugin group, so for most applications, this should not be a breaking change.