Add `RigidBodyDisabled` (#536)
# Objective

Fixes part of #436.

A pretty common user request is to be able to temporarily disable rigid bodies without removing them from the world. This can be done by removing `RigidBody`, but this loses the type of rigid body, makes queries a bit more annoying, and can involve somewhat expensive clean-up.

In the absence of general-purpose component disabling in Bevy, it would be useful to have a `RigidBodyDisabled` marker that can be used to temporarily remove a body from the simulation.

## Solution

Add a `RigidBodyDisabled` marker component. It disables velocity, forces, contact response, and attached joints. It does *not* disable collision detection or spatial queries for attached colliders however; that should be handled with a separate `ColliderDisabled` component in a follow-up.