Add `ColliderDisabled` (#584)
# Objective

Closes #436.
Follow-up to #536.

Avian already supports disabling rigid bodies (#536) and joints (#519). The missing piece is disabling colliders!

## Solution

Add a `ColliderDisabled` component for temporarily disabling collision detection for colliders and removing them from spatial queries.

For now, I decided *not* to prevent disabled colliders from contributing to mass properties of rigid bodies, since the primary purpose of the component is to temporarily disable collision detection, and I think it would be unexpected if it affected the behavior of the rigid body in any other way.