Fix physics being run with zero delta time (#622)
# Objective

Fixes #619.

When pausing physics, the `PhysicsSchedule` can still run even though the delta time is zero. This can cause simulation problems and sometimes even results in values becoming NaN or infinite due to division by zero.

This is a regression from 0.1, which handled scheduling differently and did not run physics with a delta time of zero.

## Solution

Don't run the `PhysicsSchedule` if the delta time is zero.