Fix `SleepingPlugin` not being optional and refactor logic (#624)
# Objective

Avian 0.2.0 has a regression from 0.1 where the `TimeSleeping` component is required, but it is only added automatically (as a required component) if the `SleepingPlugin` is enabled. This means that disabling the `SleepingPlugin` breaks physics.

## Solution

Make `TimeSleeping` properly optional for rigid bodies.

I also added a `WakeUpBody` `Command` to clean up some logic and abstract away resetting `TimeSleeping`.