Per-manifold material properties and tangent velocity (#660)
# Objective

Follow-up to #610.

Now that we have contact modification hooks, we should add support for modifying the surface properties of contacts. This can enable use cases like conveyor belts or non-uniform friction and restitution for different parts of a mesh.

## Solution

Add `friction`, `restitution`, and `tangent_speed`/`tangent_velocity` properties to `ContactManifold`. These are copied over for each `ContactConstraint`.

> [!NOTE]
> **Why not have a hook for modifying `ContactConstraint` directly instead?**
>
> While we could technically do this right now, once we have wide SIMD constraints, `ContactConstraint` will also have its own SIMD version that is not as user-friendly and depends on feature flags. Providing hooks for modifying that is not very viable, and it would not be good for UX. Splitting contact modification into two hooks for different types of contact types could also be confusing for users.

The new *tangent velocity* is a way to emulate relative movement of contact surfaces, which is often used for conveyor belts. A new `conveyor_belt` example has been added to demonstrate this.

https://github.com/user-attachments/assets/3f07b4f1-6ff9-4b00-80db-5500ba43940b

Engines like Rapier, Jolt, and Box2D also have tangent velocity and other material properties, but the capabilities for contact modification are different.

- **Rapier**: Friction, restitution, and tangent velocity can be set for each solver contact point.
- **Jolt**: Friction, restitution, and tangent velocity can be set for each contact manifold.
- **Box2D**: Friction, restitution, and tangent velocity cannot be modified in pre-solve hooks, *but* friction and restitution have their own optional callbacks for basic mixing logic (I believe Jolt also has this), and chain shapes support per-segment materials. So, surface properties can only be set per shape (pair), but it is still decently flexible for 2D.

I opted for Jolt's approach of per-manifold configuration for now. It lets you have different material properties for e.g. different triangles of a trimesh, while keeping memory usage fairly minimal by not storing the properties for each individual contact point. I think this is a good balance of flexibility and cost.

If important use cases that require per-point surface properties arise, we could switch to that, but I'd prefer to keep things more simple and minimal for now.