Change `ColliderParent` to a `ColliderOf` relationship and rework `ColliderHierarchyPlugin` (#671)
# Objective

`ColliderParent` should be a relationship for robustness and semantics, and to take advantage of automatic source tracking for rigid bodies.

The `ColliderHierarchyPlugin` also currently deals with both attaching colliders to rigid bodies and managing transforms. This would be good to split into separate plugins.

## Solution

Rename `ColliderParent` to `ColliderOf`, make it a `Relationship`, and add a `RigidBodyColliders` component for the `RelationshipTarget`. Collider hierarchies are now managed via observers and hooks instead of systems.

Note that `ColliderOf` requires a custom `Relationship` implementation, because Bevy's default implementation does not allow relationships to point to themselves.

I also split the `ColliderHierarchyPlugin` into a `ColliderHierarchyPlugin` (manages `ColliderOf` relationships) and `ColliderTransformPlugin` (manages and propagates collider transforms).

---

## Migration Guide

`ColliderParent` has been renamed to `ColliderOf`, and it is now a `Relationship`.

The transform management in `ColliderHierarchyPlugin` has been extracted into a new `ColliderTransformPlugin`. The `ColliderHierarchyPlugin` no longer takes a schedule.