Rename `Contacts` to `ContactPair` (#685)
# Objective

Especially after #683, we use the term "contact pair" a lot. However, the contact pair type is called `Contacts`. This name is very ambiguous: does it represent all contacts in the world, all contacts between two entities, contacts belonging to a specific contact surface, or something else?

## Solution

Rename `Contacts` to `ContactPair`. It much more accurately describes what the type represents: contact data for a pair of entities that may be in contact. This is also the name used by Rapier.

---

## Migration Guide

`Contacts` has been renamed to `ContactPair`.