Create thread-local contact status bit vectors with the correct block count (#686)
# Objective

Fixes #687.

The narrow phase is currently initializing the thread-local contact status bit vectors with a block count of zero. This means that for the first tick, contact status changes might not be tracked correctly. The bitwise OR to combine the bit vectors also currently panics with debug assertions enabled, because the sizes of the vectors don't match.

## Solution

Initialize thread-local contact status bit vectors with the correct block count.

I also improved the error message for the debug assertion in `BitVec::or`.