Fix removing colliders from contact graph `EntityDataIndex` (#688)
# Objective

When a collider is removed from the contact graph, the `EntityDataIndex` used for mapping entities to graph nodes is not updated. This can cause bugs and even crashes when an entity with the same index is removed and later added back to the graph.

## Solution

Remove colliders from the `EntityDataIndex` in `ContactGraph::remove_collider_with`.