Fix contact pair removal bugs (#693)
# Objective

Contact pair removal currently has some issues.

1. Contact pair removal is done using the `EdgeIndex` directly, but removing an edge can move the indices around (due to `swap_remove`), which can sometimes result in incorrect pair removal if many pairs are removed at once.
2. Changing `CollisionLayers` doesn't update existing contact pairs.
3. When a collider is removed or disabled, `CollisionEnded` is sent even for non-touching contact pairs.

## Solution

Fix the things!

For problem 2, I just made the narrow phase always check if the `CollisionLayers` match for now. Ideally, we'd handle this with fine-grained change detection instead.

`CollisionLayers` is now also a required component for colliders. This avoids extra branching and is perhaps more semantically accurate; colliders can only belong to collision layers if they have the component for it.

---

## Migration Guide

`CollisionLayers` is now a required component for colliders.