Reduce `SpatialQueryPipeline` overhead (#696)
# Objective

The `SpatialQueryPipeline` currently has some egregious overhead for scenes with a large number of entities. It allocates an `EntityHashMap` with *all* colliders (`Arc`ed) and some other data, from scratch, *every frame*. Even worse, it allocates QBVH nodes for all entities up to the largest collider `Entity` index. For a scene with 999,999 non-physics entities and one collider, it seems to allocate a *million* QBVH nodes every frame. This produces profiles like this:

![Really bad spatial query pipeline performance](https://github.com/user-attachments/assets/802553c1-d013-4e8a-8344-e54d1e1de0b4)

Yikes!

## Solution

Just store QBVH proxies in a `Vec`, and use those vec indices rather than `Entity` indices for QBVH building. This also lets us replace a bunch of hashmap lookups with vector indexing, get rid of some `Entity` conversions, and avoid table scans for `Added<Collider>`.

This is still kind of a duct-tape fix. We should ideally have separate BVHs for dynamic or kinematic and static or sleeping bodies, along with incremental updates. It would also be good if we didn't have to store all the duplicate collider data and just queried the ECS directly, assuming the query overhead is negligible. This will be worked on in the future.

Still, this seems to be much better!

![Reduced overhead](https://github.com/user-attachments/assets/c680dd25-e4d2-4faf-9faa-002f39803c4a)