Clean up collision detection module structure and re-exports (#698)
# Objective

Some of the collision detection modules are not located in the most appropriate place, and the imports are a bit messy.

## Solution

Move

- `layers` from `collision` to `collision::collider`
- `contact_query` from `collision` to `collision::collider::parry`
- `feature_id` from `collision` to `collision::contact_types`

I also removed the `collider::*` and `layers::*` re-exports from the `collision` module, and instead added a `prelude` for collision detection functionality.

---

## Migration Guide

Some collision detection modules and imports have been reorganized.

The following modules have been moved:

- `layers` from `collision` to `collision::collider`
- `contact_query` from `collision` to `collision::collider::parry`
- `feature_id` from `collision` to `collision::contact_types`

Previously, a lot of collision detection types were also re-exported directly from the `collision` module. Now, there is instead a `prelude` for the `collision` module.