Sort contact constraints for determinism when `parallel` feature is enabled (#712)
# Objective

Multithreaded determinism is currently broken! This is because #699 accidentally resulted in contact constraint order being non-deterministic. CI didn't catch this, because it currently doesn't use the `parallel` feature.

## Solution

When the `parallel` feature is enabled, store the pair index for each `ContactConstraint`, and just sort the constraints based on these indices.

I suspect that a better approach that doesn't require sorting exists, but for now, this works as a simple hot-fix. I also tried the simple approach of returning `Vec`s from Bevy's `par_splat_map_mut`, but that would mean that we can't reuse the constraint buffers, and have to allocate from scratch every time. Either way, there's more to try and experiment with here.

Because the constraints within chunks are still largely sorted, the sorting ends up being quite fast. In the `pyramid_2d` example, it increased the time of "Update Contacts" from about 0.42 ms to about 0.44 ms with ~3025 constraints.

I also enabled the `parallel` feature for CI to hopefully catch multithreading problems like this automatically in the future!