Rename most `rigid_body` names to `body` (#718)
# Objective

We have a lot of properties called `rigid_body`. In a lot of cases, really the relevant part is that it is just a physics "body", and the `rigid_` qualifier is largely unnecessary. Typically, it's enough to just use `body` for properties, and it can result in cleaner names.

## Solution

Rename `rigid_body` to just `body` for `OnCollisionStart`, `OnCollisionEnd`, `ColliderOf`, and many internal variables and types.

Despite using `body` for a lot of property names, I think we should still probably keep the `Rigid` prefix for types such as `RigidBody`.

---

## Migration Guide

`ColliderQuery::rigid_body` has been renamed to `of` (for `ColliderOf`) and there is a new `body` helper to get the contained entity directly.