Fix `body` property of collision events and reflect `CollisionEventsEnabled` properly (#752)
# Objective

Fixes #750

Someone (me) somehow got the entities mixed up for the `body` property of the `OnCollisionStart` and `OnCollisionEnded` events. It is documented to be the body of the entity that hit `trigger.target()`, but it is in fact the body associated with the `trigger.target()` itself. Yikes!

Additionally, the `CollisionEventsEnabled` type doesn't reflect `Component` and isn't registered. This breaks scenes.

## Solution

Fix `body` to be the body of the entity that hit `trigger.target()`, and reflect `Component` for `CollisionEventsEnabled`.

This is critical enough that I'll make a 0.3.1 patch release with it soon.