Fix AABB interval initialization in broad phase (#772)
# Objective

Fixes #739.

The broad phase uses `RemovedComponents` for detecting when `Disabled` or `ColliderDisabled` is removed. However, it can miss removals in schedules with fixed time steps. To make it more robust, we should use observers instead.

Additionally, there is a bug where child colliders of static bodies are not marked as static/inactive on the first time step.

## Solution

Use observers to fix the first problem, and rework initialization logic to fix the second.